requirements.txt
asgiref==3.9.1
Django==5.2.6
django-filter==25.1
django-stubs==5.2.5
django-stubs-ext==5.2.5
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
graphql-core==3.2.6
gunicorn==23.0.0
lia-web==0.2.3
mypy==1.18.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
six==1.17.0
sqlparse==0.5.3
strawberry-django-auth==0.378.4
strawberry-graphql==0.282.0
strawberry-graphql-django==0.65.1
types-PyYAML==6.0.12.20250915
typing_extensions==4.15.0
whitenoise==6.11.0
